implemented PETAcquisitionModel.norm(), addresses #804
