--- src/xSTIR/cSTIR/include/sirf/STIR/bf_operator.h
@@ -1,30 +1,0 @@
-#pragma once
-
-#include "sirf/STIR/stir_x.h"
-
-namespace sirf {
-	class BFOperator : public anOperator<STIRImageData> {
-	public:
-		BFOperator(const PETAcquisitionModel& am) : sptr_am_(am.linear_acq_mod_sptr()) {}
-		void set_subset(int sub_num)
-		{
-			sub_num_ = sub_num;
-		}
-		void set_num_subsets(int num_sub)
-		{
-			num_sub_ = num_sub;
-		}
-		virtual std::shared_ptr<STIRImageData> apply(const STIRImageData& image_data) const
-		{
-			std::shared_ptr<PETAcquisitionData> sptr_fwd = 
-				sptr_am_->forward(image_data, sub_num_, num_sub_); // , true);
-			std::shared_ptr<STIRImageData> sptr_bwd = 
-				sptr_am_->backward(*sptr_fwd, sub_num_, num_sub_);
-			return sptr_bwd;
-		}
-	private:
-		std::shared_ptr<PETAcquisitionModel> sptr_am_;
-		int sub_num_ = 0;
-		int num_sub_ = 1;
-	};
-}
